requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_pytorch_select=1.1.0=cpu
alabaster=0.7.12=py37_0
argh=0.26.2=py37_0
asn1crypto=0.24.0=py37_0
astroid=2.2.5=py37_0
atari-py=0.1.15=pypi_0
atomicwrites=1.3.0=py37_1
attrs=19.1.0=py_0
autopep8=1.4.4=py_0
babel=2.7.0=py_0
backcall=0.1.0=py37_0
bcrypt=3.1.7=py37he774522_0
blas=1.0=mkl
bleach=3.1.0=py37_0
ca-certificates=2020.4.5.1=hecc5488_0
certifi=2020.4.5.1=py37hc8dfbb8_0
cffi=1.12.3=py37h7a1dbc1_0
chardet=3.0.4=py37_1003
cloudpickle=1.0.0=py_0
colorama=0.4.1=py37_0
cryptography=2.7=py37h7a1dbc1_0
cudatoolkit=10.1.168=0
cycler=0.10.0=py37_0
cython=0.29.13=py37ha925a31_0
decorator=4.4.0=py_0
deepdish=0.3.6=pypi_0
defusedxml=0.6.0=py_0
diff-match-patch=20181111=py_0
docutils=0.15.2=py37_0
entrypoints=0.3=py37_0
ffmpeg=4.2.3=ha925a31_0
flake8=3.7.9=py37_0
freetype=2.9.1=ha9979f8_1
future=0.17.1=pypi_0
git=2.20.1=h6bb4b03_0
glfw=1.8.1=pypi_0
gym=0.12.4=pypi_0
html5lib=1.0.1=py_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
idna=2.8=py37_0
imageio=2.5.0=pypi_0
imagesize=1.1.0=py37_0
importlib_metadata=1.4.0=py37_0
intel-openmp=2019.3=203
intervaltree=3.0.2=py_0
ipykernel=5.1.4=py37h39e3cac_0
ipython=7.8.0=py37h39e3cac_0
ipython_genutils=0.2.0=py37_0
isort=4.3.21=py37_0
jedi=0.14.1=py37_0
jinja2=2.10.1=py37_0
jpeg=9b=hb83a4c4_2
jsonschema=3.0.2=py37_0
jupyter_client=5.3.4=py37_0
jupyter_core=4.6.1=py37_0
keyring=18.0.0=py37_0
kiwisolver=1.1.0=py37ha925a31_0
lazy-object-proxy=1.4.2=py37he774522_0
libpng=1.6.37=h2a8f88b_0
libpython=2.1=py37_0
libsodium=1.0.16=h9d3ae62_0
libspatialindex=1.9.3=h33f27b4_0
libtiff=4.0.10=hb898794_2
lockfile=0.12.2=pypi_0
m2w64-binutils=2.25.1=5
m2w64-bzip2=1.0.6=6
m2w64-crt-git=5.0.0.4636.2595836=2
m2w64-gcc=5.3.0=6
m2w64-gcc-ada=5.3.0=6
m2w64-gcc-fortran=5.3.0=6
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gcc-objc=5.3.0=6
m2w64-gmp=6.1.0=2
m2w64-headers-git=5.0.0.4636.c0ad18a=2
m2w64-isl=0.16.1=2
m2w64-libiconv=1.14=6
m2w64-libmangle-git=5.0.0.4509.2e5a9a2=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
m2w64-make=4.1.2351.a80a8b8=2
m2w64-mpc=1.0.3=3
m2w64-mpfr=3.1.4=4
m2w64-pkg-config=0.29.1=2
m2w64-toolchain=5.3.0=7
m2w64-tools-git=5.0.0.4592.90b8472=2
m2w64-windows-default-manifest=6.4=3
m2w64-winpthreads-git=5.0.0.4634.697f757=2
m2w64-zlib=1.2.8=10
markupsafe=1.1.1=py37he774522_0
matplotlib=3.1.1=py37hc8f65d3_0
matplotlib-label-lines=0.3.6=pypi_0
mccabe=0.6.1=py37_1
mistune=0.8.4=py37he774522_0
mkl=2019.4=245
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.0.14=py37h14836fe_0
mkl_random=1.1.0=py37h675688f_0
mock=3.0.5=pypi_0
more-itertools=8.0.2=py_0
msys2-conda-epoch=20160418=1
nbconvert=5.5.0=py_0
nbformat=4.4.0=py37_0
ninja=1.9.0=py37h74a9793_0
numexpr=2.6.9=pypi_0
numpy=1.16.5=py37h19fb1c0_0
numpy-base=1.16.5=py37hc3f5095_0
numpydoc=0.9.1=py_0
olefile=0.46=py37_0
opencv-python=4.1.0.25=pypi_0
openssl=1.1.1g=he774522_0
packaging=19.2=py_0
pandas=0.25.1=py37ha925a31_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py37_1
paramiko=2.6.0=py37_0
parso=0.6.0=py_0
pathtools=0.1.2=py_1
patsy=0.5.1=py37_0
pexpect=4.8.0=py37_0
pickleshare=0.7.5=py37_0
pillow=6.0.0=pypi_0
pip=19.2.3=py37_0
pluggy=0.13.1=py37_0
prompt_toolkit=2.0.9=py37_0
psutil=5.6.3=py37he774522_0
pycodestyle=2.5.0=py37_0
pycparser=2.19=py37_0
pydocstyle=4.0.1=py_0
pyflakes=2.1.1=py37_0
pygame=1.9.6=pypi_0
pyglet=1.3.2=pypi_0
pygments=2.4.0=py_0
pylint=2.3.1=py37_0
pympler=0.7=py_0
pynacl=1.3.0=py37h62dcd97_0
pyopenssl=19.0.0=py37_0
pyparsing=2.4.0=py_0
pyqt=5.9.2=py37h6538335_2
pyrsistent=0.15.4=py37he774522_0
pysocks=1.7.1=py37_0
pystan=2.19.0.0=py37hcce6980_0
python=3.7.3=h8c8aaf0_1
python-dateutil=2.8.0=py37_0
python-jsonrpc-server=0.3.4=py_0
python-language-server=0.31.7=py37_0
python_abi=3.7=1_cp37m
pytorch=1.0.0=py3.7_cuda90_cudnn7_1
pytz=2019.1=py_0
pywin32=223=py37hfa6e2cd_1
pyyaml=5.2=py37he774522_0
pyzmq=18.1.0=py37ha925a31_0
qdarkstyle=2.8=py_0
qt=5.9.7=vc14h73c81de_0
qtawesome=0.6.0=py_0
qtconsole=4.6.0=py_1
qtpy=1.7.1=py_0
requests=2.22.0=py37_0
rope=0.14.0=py_0
rtree=0.9.3=py37h21ff451_0
scipy=1.3.0=pypi_0
seaborn=0.9.0=py37_0
setuptools=41.2.0=py37_0
simplegeneric=0.8.1=py37_2
sip=4.19.8=py37h6538335_0
six=1.12.0=py37_0
snowballstemmer=1.9.1=py_0
sortedcontainers=2.1.0=py37_0
sphinx=2.2.0=py_0
sphinxcontrib=1.0=py37_1
sphinxcontrib-applehelp=1.0.1=py_0
sphinxcontrib-devhelp=1.0.1=py_0
sphinxcontrib-htmlhelp=1.0.2=py_0
sphinxcontrib-jsmath=1.0.1=py_0
sphinxcontrib-qthelp=1.0.2=py_0
sphinxcontrib-serializinghtml=1.1.3=py_0
sphinxcontrib-websupport=1.1.2=py_0
spyder=4.0.0=py37_0
spyder-kernels=1.8.1=py37_0
sqlite=3.28.0=he774522_0
statsmodels=0.10.1=py37h8c2d366_0
swig=3.0.12=h047fa9f_3
tables=3.5.1=pypi_0
testpath=0.4.2=py37_0
tk=8.6.8=hfa6e2cd_0
torchvision=0.2.2=py_3
tornado=6.0.3=py37he774522_0
traitlets=4.3.2=py37_0
ujson=1.35=py37hfa6e2cd_0
urllib3=1.24.2=py37_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_0
watchdog=0.9.0=py37_1
wcwidth=0.1.7=py37_0
webencodings=0.5.1=py37_1
wheel=0.33.6=py37_0
win_inet_pton=1.1.0=py37_0
win_unicode_console=0.5=py37_0
wincertstore=0.2=py37_0
wrapt=1.11.2=py37he774522_0
xz=5.2.4=h2fa13f4_4
yaml=0.1.7=hc54c509_2
yapf=0.28.0=py_0
zeromq=4.3.1=h33f27b4_3
zipp=0.6.0=py_0
zlib=1.2.11=h62dcd97_3
zstd=1.3.7=h508b16e_0
